refactor(linter): move diagnostic service to oxc_linter
