+++ crates/oxc_diagnostics/Cargo.toml
@@ -19,5 +19,4 @@ workspace = true
 doctest = false
 
 [dependencies]
-cow-utils = { workspace = true }
 miette = { workspace = true }
